Bug fix
Sometimes when building the Fock matrix over with closed or virtual indices, for certain terms no valid contraction would be found.
In these cases, oldspin is a null pointer, and we can have segmentation faults when trying to generate code.
Now it will just skip that contribution in such cases.

This bug does not show up when using only general or active indices, which is apparently why it was silent in previous uses of SMITH3.
With this change, I can get the same results for MS-RelCASPT2 using either the full (g, g) Fock matrix or building each block separately.